requirements.txt
auto-editor==27.0.0
openai-whisper==20231117
ttkbootstrap==1.10.1
streamlit==1.32.0
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.1
faster-whisper==0.10.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0 